requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.2
Werkzeug==3.0.3
email_validator==2.2.0
python-dotenv==1.1.0
mysqlclient==2.2.1
PyMySQL==1.1.0